+++ view/tws_administration_view/lib/view/articles/locations/adapters/locations_whisper_options_adapter.dart
@@ -1,4 +1,4 @@
-part of '../../whispers/locations_create_whisper.dart';
+part of '../whispers/locations_create_whisper.dart';
 
 final SessionStorage _sessionStorage = SessionStorage.i;