+++ view/tws_administration_view/lib/view/articles/sections/sections_article.dart
@@ -2,8 +2,6 @@ import 'package:csm_client/csm_client.dart';
 import 'package:csm_view/csm_view.dart' hide JObject; 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:tws_administration_view/core/constants/twsa_colors.dart';
-import 'package:tws_administration_view/core/constants/twsa_common_displays.dart';
 import 'package:tws_administration_view/core/router/twsa_routes.dart';
 import 'package:tws_administration_view/data/services/sources.dart';
 import 'package:tws_administration_view/data/storages/session_storage.dart';
@@ -17,66 +15,56 @@ import 'package:tws_administration_view/view/widgets/tws_article_table/tws_artic
 import 'package:tws_administration_view/view/widgets/tws_article_table/tws_article_table_field_options.dart';
 import 'package:tws_administration_view/view/widgets/tws_autocomplete_field/tws_autocomplete_adapter.dart';
 import 'package:tws_administration_view/view/widgets/tws_autocomplete_field/tws_autocomplete_field.dart';
-import 'package:tws_administration_view/view/widgets/tws_cascade_section.dart';
 import 'package:tws_administration_view/view/widgets/tws_confirmation_dialog.dart';
-import 'package:tws_administration_view/view/widgets/tws_display_flat.dart';
 import 'package:tws_administration_view/view/widgets/tws_input_text.dart';
 import 'package:tws_administration_view/view/widgets/tws_property_viewer.dart';
-import 'package:tws_administration_view/view/widgets/tws_section.dart';
-import 'package:tws_administration_view/view/widgets/tws_section_divider.dart';
 import 'package:tws_foundation_client/tws_foundation_client.dart';
 
-part 'adapters/location_table_adapter.dart';
+part 'adapters/sections_table_adapter.dart';
 
 
-class LocationsArticle extends CSMPageBase {
+class SectionsArticle extends CSMPageBase {
   static final TWSArticleTableAgent tableAgent = TWSArticleTableAgent();
-  const LocationsArticle({super.key});
+  const SectionsArticle({super.key});
 
   @override
   Widget compose(BuildContext ctx, Size window) {
     return BusinessFrame(
-      currentRoute: TWSARoutes.locationsArticle,
+      currentRoute: TWSARoutes.sectionsArticle,
       actionsOptions: ActionRibbonOptions(
         refresher: tableAgent.refresh,
         maintenanceGroupConfig: MaintenanceGroupOptions(
-          onCreate: () => CSMRouter.i.drive(TWSARoutes.locationsCreateWhisper),
+          onCreate: () => CSMRouter.i.drive(TWSARoutes.sectionsCreateWhisper),
         ),
       ),
-      article: TWSArticleTable<Location>(
+      article: TWSArticleTable<Section>(
         editable: true,
         removable: false,
         adapter: const _TableAdapter(),
         agent: tableAgent,
-        fields: <TWSArticleTableFieldOptions<Location>>[
-          TWSArticleTableFieldOptions<Location>(
+        fields: <TWSArticleTableFieldOptions<Section>>[
+          TWSArticleTableFieldOptions<Section>(
             'Name',
-            (Location item, int index, BuildContext ctx) {
+            (Section item, int index, BuildContext ctx) {
              return item.name;
             },
           ),
-          TWSArticleTableFieldOptions<Location>(
-            'Country',
-            (Location item, int index, BuildContext ctx) {
-             return item.addressNavigation?.country ?? "---";
+          TWSArticleTableFieldOptions<Section>(
+            'Location',
+            (Section item, int index, BuildContext ctx) {
+             return item.locationNavigation?.name ?? '---';
             },
           ),
-          TWSArticleTableFieldOptions<Location>(
-            'City',
-            (Location item, int index, BuildContext ctx) {
-             return item.addressNavigation?.city ?? "---";
+          TWSArticleTableFieldOptions<Section>(
+            'Capacity',
+            (Section item, int index, BuildContext ctx) {
+             return item.capacity.toString();
             },
           ),
-          TWSArticleTableFieldOptions<Location>(
-            'Street',
-            (Location item, int index, BuildContext ctx) {
-             return item.addressNavigation?.street ?? "---";
-            },
-          ),
-          TWSArticleTableFieldOptions<Location>(
-            'Colonia',
-            (Location item, int index, BuildContext ctx) {
-             return item.addressNavigation?.colonia ?? "---";
+          TWSArticleTableFieldOptions<Section>(
+            'Ocupancy',
+            (Section item, int index, BuildContext ctx) {
+             return item.ocupancy.toString();
             },
           ),
         ],