+++ view/tws_administration_view/lib/view/pages/business/business_frame.dart
@@ -60,6 +60,16 @@ class BusinessFrame extends StatelessWidget {
       route: TWSARoutes.locationsArticle,
       title: 'Locations',
     ),
+    ArticleOptions(
+      icon: (Color? stateColor) {
+        return Icon(
+          Icons.space_dashboard_sharp,
+          color: stateColor,
+        );
+      },
+      route: TWSARoutes.sectionsArticle,
+      title: 'Sections',
+    ),
   ];
 
   @override